requirements.txt
flask
pillow
python-dotenv
qrcode
